Extracted sync methods from AdminXComponent to StateBridge service (#25292)
ref https://linear.app/ghost/issue/BER-2929

- we want to make these state syncing methods available to our in-development React admin shell which means we won't be able to pass them through as component props
- added new `state-bridge` service and moved the methods there to make them easier to expose later on
  - this is a straight code shuffle, no functional changes have been made
